+++ lib/presentation/login/login_screen.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:frontend/config/theme/color_data.dart';
+import 'package:frontend/di/getx_binding_builder_call_back.dart';
 import 'package:frontend/presentation/login/components/kakao_login_widget.dart';
 import 'package:frontend/core/utils/utils.dart';
 import 'package:frontend/presentation/login/login_view_model.dart';
@@ -19,6 +20,8 @@ class LoginScreen extends GetView<LoginViewModel> {
 
   @override
   Widget build(BuildContext context) {
+    getLoginBinding();
+
     if (isSignup) {
       controller.signupAndLogin();
     }
