+++ lib/presentation/on_boarding/on_boarding_finish/on_boarding_finish_viewmodel.dart
@@ -1,3 +1,23 @@
+import 'package:flutter_local_notifications/flutter_local_notifications.dart';
+import 'package:frontend/domain/use_case/push_message/push_message_use_case.dart';
+import 'package:frontend/main_view_model.dart';
 import 'package:get/get.dart';
 
-class OnBoardingFinishViewModel extends GetxController {}
+class OnBoardingFinishViewModel extends GetxController {
+  final PushMessageUseCase pushMessagePermissionUseCase;
+
+  OnBoardingFinishViewModel({
+    required this.pushMessagePermissionUseCase,
+  });
+
+  final mainViewController = Get.find<MainViewModel>();
+
+  setPushMessage() {
+    pushMessagePermissionUseCase.setPushMessagePermission(true.toString());
+    mainViewController.pushMessagePermission.value = true;
+
+    pushMessagePermissionUseCase.dailyAtTimeNotification(
+      const Time(21, 0, 0),
+    );
+  }
+}
