+++ lib/presentation/profile/push_message/push_message_screen.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_local_notifications/flutter_local_notifications.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'package:flutter_svg/svg.dart';
 import 'package:flutter_switch/flutter_switch.dart';
 import 'package:frontend/config/theme/color_data.dart';
 import 'package:frontend/config/theme/text_data.dart';
@@ -12,11 +11,9 @@ import 'package:frontend/core/utils/library/date_time_spinner/i18n_model.dart';
 import 'package:frontend/di/getx_binding_builder_call_back.dart';
 import 'package:frontend/main_view_model.dart';
 import 'package:frontend/presentation/components/back_icon.dart';
-import 'package:frontend/presentation/components/bottom_button.dart';
 import 'package:frontend/presentation/components/toast.dart';
 import 'package:frontend/presentation/profile/components/profile_button.dart';
 import 'package:frontend/presentation/profile/push_message/push_message_view_model.dart';
-import 'package:frontend/presentation/profile/terms/marketing_consent_screen.dart';
 import 'package:get/get.dart';
 import 'package:intl/intl.dart';
 import 'package:notification_permissions/notification_permissions.dart';
@@ -78,83 +75,7 @@ class PushMessageScreen extends GetView<PushMessageViewModel> {
                       permissionStatus = NotificationPermissions
                           .requestNotificationPermissions();
                     } else {
-                      if (!mainViewController.marketingConsentAgree.value &&
-                          !mainViewController.pushMessagePermission.value) {
-                        // ignore: use_build_context_synchronously
-                        showModalBottomSheet(
-                          shape: const RoundedRectangleBorder(
-                            borderRadius: BorderRadius.vertical(
-                              top: Radius.circular(25.0),
-                            ),
-                          ),
-                          context: context,
-                          isScrollControlled: true,
-                          builder: (context) => Padding(
-                            padding: EdgeInsets.only(
-                                bottom:
-                                    MediaQuery.of(context).viewInsets.bottom),
-                            child: Column(
-                              crossAxisAlignment: CrossAxisAlignment.center,
-                              mainAxisSize: MainAxisSize.min,
-                              children: [
-                                Align(
-                                  alignment: Alignment.centerRight,
-                                  child: Padding(
-                                    padding: EdgeInsets.only(
-                                        top: 10.0.h, right: 18.0.w),
-                                    child: IconButton(
-                                      onPressed: () {
-                                        Get.back();
-                                      },
-                                      icon: SvgPicture.asset(
-                                        "lib/config/assets/images/profile/close.svg",
-                                      ),
-                                      color: kGrayColor500,
-                                    ),
-                                  ),
-                                ),
-                                Text(
-                                  "마케팅 정보 수신동의",
-                                  style: kHeader4Style.copyWith(
-                                      color: Theme.of(context)
-                                          .colorScheme
-                                          .textTitle),
-                                ),
-                                Text(
-                                  "마케팅 정보에 수신동의를 하셔야",
-                                  style: kHeader6Style.copyWith(
-                                      color: Theme.of(context)
-                                          .colorScheme
-                                          .textSubtitle),
-                                ),
-                                Text(
-                                  "푸시메시지를 받을 수 있어요",
-                                  style: kHeader6Style.copyWith(
-                                      color: Theme.of(context)
-                                          .colorScheme
-                                          .textSubtitle),
-                                ),
-                                SizedBox(
-                                  height: 28.0.h,
-                                ),
-                                BottomButton(
-                                  title: '수신 동의',
-                                  onTap: () {
-                                    Get.back();
-                                    Get.to(
-                                      () => const MarketingConsentScreen(
-                                        isProfileScreen: true,
-                                      ),
-                                    );
-                                  },
-                                ),
-                              ],
-                            ),
-                          ),
-                        );
-                      } else {
-                        mainViewController.togglePushMessageValue();
-                      }
+                      mainViewController.togglePushMessageValue();
                     }
                   },
                 ),
